+++ oauth2_http/javatests/com/google/auth/oauth2/ImpersonatedCredentialsTest.java
@@ -530,9 +530,9 @@ public class ImpersonatedCredentialsTest extends BaseSerializationTest {
     ServiceAccountCredentials sourceCredentialsSSJ =
         ((ServiceAccountCredentials) sourceCredentials)
             .toBuilder()
-            .setHttpTransportFactory(mockTransportFactory)
-            .setUseJwtAccessWithScope(true)
-            .build();
+                .setHttpTransportFactory(mockTransportFactory)
+                .setUseJwtAccessWithScope(true)
+                .build();
     ImpersonatedCredentials targetCredentials =
         ImpersonatedCredentials.create(
             sourceCredentialsSSJ,
@@ -559,9 +559,9 @@ public class ImpersonatedCredentialsTest extends BaseSerializationTest {
     ServiceAccountCredentials sourceCredentialsNonGDU =
         ((ServiceAccountCredentials) sourceCredentials)
             .toBuilder()
-            .setUniverseDomain(TEST_UNIVERSE_DOMAIN)
-            .setHttpTransportFactory(transportFactory)
-            .build();
+                .setUniverseDomain(TEST_UNIVERSE_DOMAIN)
+                .setHttpTransportFactory(transportFactory)
+                .build();
     ImpersonatedCredentials impersonatedCredentials =
         ImpersonatedCredentials.create(
             sourceCredentialsNonGDU,
@@ -922,9 +922,9 @@ public class ImpersonatedCredentialsTest extends BaseSerializationTest {
     ServiceAccountCredentials sourceCredentialsNonGDU =
         ((ServiceAccountCredentials) sourceCredentials)
             .toBuilder()
-            .setUniverseDomain("test.com")
-            .setHttpTransportFactory(transportFactory)
-            .build();
+                .setUniverseDomain("test.com")
+                .setHttpTransportFactory(transportFactory)
+                .build();
     ImpersonatedCredentials targetCredentials =
         ImpersonatedCredentials.create(
             sourceCredentialsNonGDU,
@@ -1048,9 +1048,9 @@ public class ImpersonatedCredentialsTest extends BaseSerializationTest {
     ServiceAccountCredentials sourceCredentialsNonGDU =
         ((ServiceAccountCredentials) sourceCredentials)
             .toBuilder()
-            .setUniverseDomain("test.com")
-            .setHttpTransportFactory(transportFactory)
-            .build();
+                .setUniverseDomain("test.com")
+                .setHttpTransportFactory(transportFactory)
+                .build();
     ImpersonatedCredentials targetCredentials =
         ImpersonatedCredentials.create(
             sourceCredentialsNonGDU,